Convert into plugin
Core functionality should not be implemented as child theme of astra or even just a parent theme. As client site may use some theme already and it will be problematic to convert core functionality implemented as theme into child theme of a random parent theme. The biggest problem will be keeping ability to pull core updates in the future (keep core up to date) as core files will probably contain some custom code.
So keeping the core functionality as a plugin allows to keep core functionality up to date with less problems.